Drop all BT keyboards and mice from Android's input processing
For now, ignore all BT keyboards and mice since Maru Desktop will be
using them.

TODO: this should be dynamically switchable though so that they are only ignored
if Maru Desktop is running.

Tested with a bluetooth mouse that didn't use the "Bluetooth Mouse" device
name and it was succesfully dropped. This works in conjunction with the
excluded-input-devices.xml, with the xml being applied first.

 Conflicts:
	services/inputflinger/EventHub.cpp

Signed-off-by: NoelMacwan <yulespencer@gmail.com>